% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsHourlyPrecipitation.R
\name{awsGetHourlyPrecip}
\alias{awsGetHourlyPrecip}
\title{Get hourly precipitation data.}
\usage{
awsGetHourlyPrecip(start_time, end_time, url)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:00"}

\item{end_time}{end time, format "YYYY-mm-dd HH:00"}

\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}
}
\value{
A named list of coordinates (list name "coords"), dates (list name "date") and the data (list name "data")
        \itemize{
           \item{\strong{coordinates}: }{a data.frame with column names "id", "longitude" and "latitude"}
           \item{\strong{dates}: }{a vector of dates in the format "YYYYmmddHH" in local time}
           \item{\strong{data}: }{a matrix with row number equals to the length of dates and column number equals to the length of coordinates}
         }
}
\description{
Get hourly precipitation data for all AWS.
}
