% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsHourlyPrecipitation.R
\name{awsWriteHourlyPrecip2CDT}
\alias{awsWriteHourlyPrecip2CDT}
\title{Write hourly precipitation data to CDT format.}
\usage{
awsWriteHourlyPrecip2CDT(aws_data, csvfile)
}
\arguments{
\item{aws_data}{Hourly precipitation data output from \code{awsGetHourlyPrecip}}

\item{csvfile}{Full path to the CSV file to save the CDT format data}
}
\description{
Write hourly precipitation data to CDT format.
}
