% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQPECAPPI.R
\name{computeQPECAPPI}
\alias{computeQPECAPPI}
\title{Compute QPE from CAPPI.}
\usage{
computeQPECAPPI(
  url,
  dirOUT,
  start_time,
  end_time,
  cappi = list(method = "composite_altitude", pars = list(fun = "maximum", min_alt =
    1.7, max_alt = 15)),
  qpe = list(method = "RATE_Z", pars = list(alpha = 300, beta = 1.4, invCoef = FALSE)),
  dbz_thres = list(min = 20, max = 65),
  apply_cmd = TRUE,
  pia = NULL,
  filter = NULL,
  time_zone = "Africa/Kigali"
)
}
\arguments{
\item{url}{The URL of the server. Ex: "http://192.168.1.10:8080"}

\item{dirOUT}{Full path to the directory to save the output}

\item{start_time, end_time}{The start and end time same time zone as \code{time_zone}, format "YYYY-mm-dd HH:MM"}

\item{cappi}{A named list of parameters to be used to create the CAPPI.}

\item{qpe}{A named list of parameters to be used to compute the precipitation rate.}

\item{dbz_thres}{A named list of the minimum and maximum reflectivity threshold.
Default \code{dbz_thres = list(min = 20, max = 65)}}

\item{apply_cmd}{Logical, apply clutter mitigation decision to the fields to use to compute the QPE.
Default \code{TRUE}, applying CMD.}

\item{pia}{A named list of parameters to be used to correct the attenuation.
Default \code{NULL}, no correction performed.}

\item{filter}{A named list of parameters to be used to filter the data.
Default \code{NULL}, no filtering applied.}

\item{time_zone}{the time zone of \code{start_time}, \code{end_time} and the output QPE.
Options: "Africa/Kigali" or "UTC". Default "Africa/Kigali"}
}
\value{
A netCDF files containing the precipitation rate and accumulation
        for a single radar scan saved under the folder \code{dirOUT}
        with a file name format "precip_YYYYmmddHHMMSS.nc", same time zone as \code{time_zone}.
}
\description{
Compute QPE from CAPPI for single scan radar polar data.
}
\section{CAPPI parameters}{

cappi
}

\section{QPE parameters}{

qpe
}

\section{Attenuation correction parameters}{

pia
}

\section{Filter parameters}{

filter
}

