% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{computeValidationStats}
\alias{computeValidationStats}
\title{Compute validation statistics.}
\usage{
computeValidationStats(
  stn_data,
  grd_data,
  validation_type = "stations",
  categorical_thres = 1,
  categorical_fun = ">=",
  volumetric_type = "threshold",
  volumetric_pars = NULL
)
}
\arguments{
\item{stn_data, grd_data}{A named list of CDT stations data format, output from \code{readCDTStationFormat}.
\itemize{
\item{\strong{stn_data}: }{contains the reference observations data used to validate}
\item{\strong{grd_data}: }{contains the simulated or estimated data to be validated}
}}

\item{validation_type}{Type of the output statistics to compute. Options are
\itemize{
\item{"stations": }{The statistics will be computed for each stations}
\item{"all": }{The statistics will be computed by using all stations together}
\item{"average": }{The statistics will be computed by spatially averaging all stations}
}}

\item{categorical_thres}{A number, giving the threshold to use to create the contingency table for the categorical statistics.}

\item{categorical_fun}{A character, indicating the operator function to be applied for the categorical statistics. Options: ">=", ">", "<=", "<"}

\item{volumetric_type}{The type of threshold to use to compute the volumetric statistics. Options: "threshold", "percentile"}

\item{volumetric_pars}{The thresholds or percentile to use to compute the volumetric statistics. If \code{volumetric_type} is
\itemize{
\item{"threshold": }{\code{volumetric_pars} specifies the threshold to be used.\cr
If one value is provided, all the stations will use this threshold\cr
If a vector is provided, it must be the same number of the number of stations in \code{stn_data}
}
\item{"percentile": }{a value of the percentile to use. The percentile will be computed using \code{stn_data} 
 to create the threshold to be used to compute the volumetric statistics}
}}
}
\value{
A list object
\itemize{
\item{\strong{continuous}: }{the continuous statistics. See \code{\link{stats_continuous}} }
\item{\strong{categorical}: }{the categorical statistics. See \code{\link{stats_categorical}} }
\item{\strong{volumetric}: }{the volumetric statistics. See \code{\link{stats_volumetric_quantile}} }
\item{\strong{categorical_pars}: }{Parameters used to compute the categorical statistics.}
\item{\strong{volumetric_pars}: }{Parameters used to compute the volumetric statistics.}
}
}
\description{
A wrapper to compute continuous, categorical and volumetric statistics for validation.
}
