% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCAPPI.R
\name{createCAPPI}
\alias{createCAPPI}
\title{Create CAPPI data.}
\usage{
createCAPPI(
  url,
  source,
  dirOUT,
  start_time,
  end_time,
  fields,
  cappi = list(method = "composite_altitude", pars = list(fun = "maximum", min_alt =
    1.7, max_alt = 15)),
  apply_cmd = FALSE,
  pia = NULL,
  dbz_fields = NULL,
  filter = NULL,
  filter_fields = NULL,
  time_zone = "Africa/Kigali"
)
}
\arguments{
\item{url}{The URL of the server. Ex: "http://192.168.1.10:8080"}

\item{source}{A partial path of the directory containing the folders dates of the mdv files.
Ex: "radarPolar/ops1/sur" or "radarPolar/ops/sur"}

\item{dirOUT}{Full path to the directory to save the output}

\item{start_time, end_time}{The start and end time same time zone as \code{time_zone}, format "YYYY-mm-dd HH:MM"}

\item{fields}{Vector of the fields to create.}

\item{cappi}{A named list of parameters to be used to create the CAPPI.}

\item{apply_cmd}{Logical, apply clutter mitigation decision to \code{fields}.}

\item{pia}{A named list of of the method and parameters to use to perform an attenuation correction for the reflectivity fields \code{dbz_fields}.
Default \code{NULL}, no attenuation correction performed.}

\item{dbz_fields}{A vector of reflectivity fields to correct the attenuation. Must be in \code{fields}. Default \code{NULL}}

\item{filter}{A named list of the method and parameters to use to filter \code{filter_fields}.
Default \code{NULL}, no filtering applied.}

\item{filter_fields}{A vector of fields in which the filter will be applied. Must be in \code{fields}. Default \code{NULL}}

\item{time_zone}{the time zone of \code{start_time}, \code{end_time} and the output CAPPI  data.
Options: "Africa/Kigali" or "UTC". Default "Africa/Kigali"}
}
\description{
Create CAPPI data for single scan radar polar.
}
