% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUploadDir.R
\name{deleteUploadDir}
\alias{deleteUploadDir}
\title{Delete user upload folder}
\usage{
deleteUploadDir(dirUPLOAD)
}
\arguments{
\item{dirUPLOAD}{path to upload folder}
}
\description{
Delete user upload folder.
}
