% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipAccumulation.R
\name{down_qpe_accumulation}
\alias{down_qpe_accumulation}
\title{Precipitation accumulation}
\usage{
down_qpe_accumulation(tstep, time, accumul, min_frac, dirQPE, dirDOWN)
}
\arguments{
\item{tstep}{time step of the data: "hourly" or "daily"}

\item{time}{the time/date at which the total will be computed}

\item{accumul}{accumulation duration}

\item{min_frac}{minimum fraction of non missing values}

\item{dirQPE}{full path to the folder containing the qpe data}

\item{dirDOWN}{full path to the folder to write the NetCDF file}
}
\description{
Write precipitation accumulation to a file in NetCDF format for download.
}
