% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract3DGridData.R
\name{extract_3DGridData}
\alias{extract_3DGridData}
\title{Extract 3D grid data.}
\usage{
extract_3DGridData(grid_coords, grid_data, points, levels, padxyz, fun_sp)
}
\arguments{
\item{grid_coords}{A named list containing the coordinates of the grid data, with names "lon", "lat", "alt"}

\item{grid_data}{A named list containing the gridded data, the 3d array data are flatten to a vector. 
The names of list are the names of the data.}

\item{points}{A data frame of the points to extract. Data frame with column names "id", "longitude" and "latitude"}

\item{levels}{A vector of the index of altitudes to be extracted in integer, or -1 to extract all available altitude}

\item{padxyz}{A vector of the padding to use in number of pixels, in order "lon", "lat", "alt"}

\item{fun_sp}{Character, function to be used for the padding. Options: "mean", "median", "max", "min"}
}
\value{
A named list \cr
        List of matrix of the extracted data over the set of points. \cr
        The row represents the set of points and the column for the levels.
}
\description{
Extract 3D grid data over a set of points.
}
