% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract3DPolardata.R
\name{extract_3DPolarData}
\alias{extract_3DPolarData}
\title{Extract 3D Polar data.}
\usage{
extract_3DPolarData(
  grid_coords,
  grid_data,
  grid_interp,
  points,
  padxy,
  fun_sp,
  maxdist = 1000
)
}
\arguments{
\item{grid_coords}{A named list containing the coordinates of the grid data, with names "x", "y", "z"}

\item{grid_data}{A named list containing the gridded data, the 3d array data are flatten to a vector. 
The names of list are the names of the data.}

\item{grid_interp}{A named list containing the coordinates at which the polar data will be interpolated,
with names "x", "y", "z", "lon" and "lat"}

\item{points}{A data frame of the points to extract. Data frame with column names "id", "longitude" and "latitude"}

\item{padxy}{A vector of the padding to use in number of pixels, in order "x", "y"}

\item{fun_sp}{Character, function to be used for the padding. Options: "mean", "median", "max", "min"}

\item{maxdist}{The maximum radius of influence in meters for the interpolation.}
}
\value{
A named list
        List of matrix of the extracted data over the set of points.
        The row represents the set of points and the column for the heights.
}
\description{
Interpolate and extract 3D polar data over a set of points.
}
