% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCAPPI.R
\name{extractCAPPI}
\alias{extractCAPPI}
\title{Extract CAPPI}
\usage{
extractCAPPI(
  cappi_dir,
  start_time,
  end_time,
  fields,
  points,
  padxy = c(0, 0),
  fun_sp = "mean"
)
}
\arguments{
\item{cappi_dir}{Full path to folder containing the CAPPI netCDF files}

\item{start_time, end_time}{The start and end time same time zone as \code{time_zone}, 
single scan format "YYYY-mm-dd HH:MM", hourly format "YYYY-mm-dd HH:00"}

\item{fields}{A vector of fields to extract}

\item{points}{A data frame of the points to extract. Data frame with column names "id", "longitude" and "latitude"}

\item{padxy}{A vector of the padding to use in number of pixels, in order "lon", "lat".
Default c(0, 0), no padding applied.}

\item{fun_sp}{Character, function to be used for the padding. Options: "mean", "median", "max", "min"

 @return A data.frame}
}
\description{
Extract CAPPI for single scan over a set of points
}
