% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractionGeom.R
\name{extractGeomRect}
\alias{extractGeomRect}
\title{Geom points extraction}
\usage{
extractGeomRect(gridObj, bbox)
}
\arguments{
\item{gridObj}{sp SpatialPixels}

\item{bbox}{list(minlon = 33, maxlon = 35, minlat = -14, maxlat = -10)}
}
\value{
list
}
\description{
Geom points extraction.
}
