% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataExtractionQPE.R
\name{extractQPE}
\alias{extractQPE}
\title{Extract QPE}
\usage{
extractQPE(dirQPE, dirDOWN, timestep, timerange, geom, min_frac = 1)
}
\arguments{
\item{dirQPE}{path to qpe folder}

\item{dirDOWN}{path to save the extracted qpe}

\item{timestep}{time step of the data}

\item{timerange}{time range, list(start, end)}

\item{geom}{list, geom to extract}

\item{min_frac}{minimum fraction of non missing values, for pentad and dekadal data}
}
\value{
a JSON object
}
\description{
Extract QPE.
}
