% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractComputedQPE.R
\name{extractQPEData}
\alias{extractQPEData}
\title{Extract Computed QPE}
\usage{
extractQPEData(
  start_time,
  end_time,
  dirOUT,
  qpe_dir = "",
  qpe_format = "precip_\%s\%s\%s\%s\%s\%s.nc",
  extr_type = "points",
  points = NULL,
  padxy = c(0, 0),
  shapefiles = NULL,
  shp_attr_name = NULL,
  shp_attr_values = NULL,
  sp_average = TRUE
)
}
\arguments{
\item{start_time, end_time}{The start and end time same time zone as \code{time_zone}, 
single scan format "YYYY-mm-dd HH:MM", hourly format "YYYY-mm-dd HH:00"}

\item{dirOUT}{Full path to folder to save the extracted QPE}

\item{qpe_dir}{Full path to folder containing the QPE netCDF files}

\item{qpe_format}{File name format of the netCDF files.
For QPE from single scan, must have 6 "\%s" and 4 "\%s" for hourly data}

\item{extr_type}{The extraction support, "points" or "shapefiles"}

\item{points}{A data frame of the points to extract. Data frame with column names "id", "longitude" and "latitude"}

\item{padxy}{If \code{extr_type} is "points", a vector of the padding to use, in order "lon", "lat".
Default c(0, 0), no padding applied.}

\item{shapefiles}{If \code{extr_type} is "shapefiles", full path to the shapefiles (with the extension ".shp") to use to extract the data}

\item{shp_attr_name}{If \code{extr_type} is "shapefiles", attribute to be used to extract the data}

\item{shp_attr_values}{If \code{extr_type} is "shapefiles", vector of the values for \code{shp_attr_name} to extract.
Default NULL, extract all polygons.}

\item{sp_average}{If \code{extr_type}  is "shapefiles", if TRUE the extracted data  are spatially average over  the polygons of the shapefiles}
}
\description{
Extract Computed QPE for single scan and hourly
}
