% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRadarPolar.R
\name{extractRadarPolar}
\alias{extractRadarPolar}
\title{Compute QPE from CAPPI.}
\usage{
extractRadarPolar(
  url,
  source,
  start_time,
  end_time,
  fields,
  points,
  sweeps = -1,
  pia = NULL,
  dbz_fields = NULL,
  filter = NULL,
  filter_fields = NULL,
  apply_cmd = FALSE,
  time_zone = "Africa/Kigali"
)
}
\arguments{
\item{url}{The URL of the server. Ex: "http://192.168.1.10:8080"}

\item{source}{A partial path of the directory containing the folders dates of the mdv files.
Ex: "radarPolar/derived/sur", "radarPolar/ops1/sur" or "radarPolar/ws/sur"}

\item{start_time, end_time}{The start and end time same time zone as \code{time_zone}, format "YYYY-mm-dd HH:MM"}

\item{fields}{Vector of the fields to extract.}

\item{points}{A data frame of the points to extract. Data frame with column names "id", "longitude" and "latitude"}

\item{sweeps}{A vector of the elevation angles to be extracted in integer, or -1 to extract all available elevation angles}

\item{pia}{A named list of of the method and parameters to use to perform an attenuation correction for the reflectivity fields before extraction.
Default \code{NULL}, no attenuation correction performed.}

\item{dbz_fields}{A vector of reflectivity fields to correct the attenuation. Must be in \code{fields}. Default \code{NULL}}

\item{filter}{A named list of the method and parameters to use to filter the fields before extraction.
Default \code{NULL}, no filtering applied.}

\item{filter_fields}{A vector of fields in which the filter will be applied. Must be in \code{fields}. Default \code{NULL}}

\item{apply_cmd}{Logical, apply clutter mitigation decision to the fields. Default \code{FALSE}}

\item{time_zone}{the time zone of \code{start_time}, \code{end_time} and the output QPE.
Options: "Africa/Kigali" or "UTC". Default "Africa/Kigali"}
}
\value{
A named list
        points: the original points used to extract the data
       date: list of dates of the extracted data
       elevation_angle: list of the elevation angles of the extracted data
       data: dictionary of longitude, latitude, altitude and the fields in the form of 3d list
            dimension: (len(date) x len(elevation_angle) x len(points))
}
\description{
Compute QPE from CAPPI for single scan radar polar data.
}
