% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCDTdata.R
\name{filterCDTStationsData}
\alias{filterCDTStationsData}
\title{Filter CDT stations data.}
\usage{
filterCDTStationsData(stn_data, grd_data)
}
\arguments{
\item{stn_data, grd_data}{A named list containing the CDT stations data, output from \code{readCDTStationFormat}}
}
\value{
A list object
\itemize{
  \item{\strong{stn}: }{The filtered \code{stn_data}}
  \item{\strong{grd}: }{The filtered \code{grd_data}}
}
}
\description{
Filter two CDT stations data to match the stations and dates.
}
