% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRadarGrid.R
\name{gridExtractedTable}
\alias{gridExtractedTable}
\title{Convert to table extracted radar Cartesian data.}
\usage{
gridExtractedTable(x)
}
\arguments{
\item{x}{A named list, output from \code{extractRadarCart}.}
}
\value{
A data.frame
}
\description{
Convert to table extracted radar Cartesian data.
}
