% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interp3d_xsec}
\alias{interp3d_xsec}
\title{Radar Cartesian cross section interpolation}
\usage{
interp3d_xsec(x, y, z, v, x_g, y_g, z_g)
}
\arguments{
\item{x, y, z}{vectors giving the coordinates of the grid to be interpolated}

\item{v}{vector of the values to be interpolated (same length as x, y and z)}

\item{x_g, y_g, z_g}{vectors giving the coordinates of the new grid (same length)}
}
\value{
a vector of the interpolated values
}
\description{
Radar Cartesian cross section interpolation.
}
