% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interp_ppi_ranges_cappi}
\alias{interp_ppi_ranges_cappi}
\title{Regrid fields from pseudo CAPPI}
\usage{
interp_ppi_ranges_cappi(x, y, z, x_g, y_g)
}
\arguments{
\item{x, y}{vectors giving the coordinates of the grid to be interpolated}

\item{z}{list of vectors containing the fields to interpolate (same length as x and y)}

\item{x_g, y_g}{vectors giving the coordinates of the new grid}
}
\value{
a list of matrices containing the interpolated fields
}
\description{
Regrid fields from pseudo CAPPI
}
