% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interp_surface_grid}
\alias{interp_surface_grid}
\title{Regrid data}
\usage{
interp_surface_grid(x, y, z, x_g, y_g, edge = TRUE)
}
\arguments{
\item{x, y}{vectors giving the coordinates of the grid to be interpolated}

\item{z}{vector of the values to be interpolated (same length as x and y)}

\item{x_g, y_g}{vectors giving the coordinates of the new grid (same length)}

\item{edge}{logical, if TRUE the grid outside the range of x and y will be extrapolated using the nearest grid}
}
\value{
a vector of the regridded values
}
\description{
Regrid data using a bilinear weights.
}
