% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRadarPolar.R
\name{polarExtractedTable}
\alias{polarExtractedTable}
\title{Convert to table extracted radar polar data.}
\usage{
polarExtractedTable(x)
}
\arguments{
\item{x}{A named list, output from \code{extractRadarPolar}.}
}
\value{
A data.frame
}
\description{
Convert to table extracted radar polar data.
}
