% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipAggregate.R
\name{precip_5min2hour}
\alias{precip_5min2hour}
\title{Precipitation aggregation to hourly}
\usage{
precip_5min2hour(
  start_time,
  end_time,
  dir5MIN,
  dirOUT,
  min_frac = 0.75,
  time_zone = "Africa/Kigali",
  in_file_prefix = "qpe_",
  out_file_prefix = "precip_"
)
}
\arguments{
\item{start_time, end_time}{start and end time of the period to aggregate
same time zone as \code{time_zone}, format "YYYY-mm-dd HH:MM"}

\item{dir5MIN}{full path to the directory containing the single scan NetCDF files}

\item{dirOUT}{full path to the directory to save the aggregate data}

\item{min_frac}{minimum fraction of non missing values}

\item{time_zone}{the time zone of \code{start_time}, \code{end_time},
the input and output netCDF files.
Options: "Africa/Kigali" or "UTC". Default "Africa/Kigali"}

\item{in_file_prefix}{prefix of the input file names. Default "qpe_"}

\item{out_file_prefix}{prefix of the output file names. Default "precip_"}
}
\description{
Aggregate precipitation accumulation from single scan to hourly.
}
