% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipAggregate.R
\name{precip_daily2month}
\alias{precip_daily2month}
\title{Precipitation aggregation from daily to monthly}
\usage{
precip_daily2month(
  start_day,
  end_day,
  dirDay,
  inFormat,
  dirOUT,
  outFormat,
  min_frac = 0.9,
  long_name = "Merged AWS-Radar monthly rainfall"
)
}
\arguments{
\item{start_day, end_day}{start and end date of the period to aggregate with format YYYY-mm-dd}

\item{dirDay}{full path to the directory containing the NetCDF files}

\item{inFormat}{format of the input file name. Ex: "precip_mrg_\%s\%s\%s.nc"}

\item{dirOUT}{full path to the directory to save the aggregate data}

\item{outFormat}{format of the output file name. Ex: "precip_mrg_\%s\%s.nc"}

\item{min_frac}{minimum fraction of non missing values}

\item{long_name}{long name for the NetCDF variable precip}
}
\description{
Aggregate daily precipitation to monthly.
}
