% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipAggregate.R
\name{precip_hour2daily}
\alias{precip_hour2daily}
\title{Precipitation aggregation from hourly to daily}
\usage{
precip_hour2daily(
  start_hour,
  end_hour,
  dirHour,
  inFormat,
  dirOUT,
  outFormat,
  obs.hour = 8,
  min.frac = 0.9,
  longname = "Merged AWS-Radar daily rainfall"
)
}
\arguments{
\item{start_hour, end_hour}{start and end time of the period to aggregate with format YYYY-mm-dd-HH in local time}

\item{dirHour}{full path to the directory containing the NetCDF files, filename time in "UTC"}

\item{inFormat}{format of the input file name. Ex: "precip_adj_\%s\%s\%s\%s.nc"}

\item{dirOUT}{full path to the directory to save the aggregate data}

\item{outFormat}{format of the output file name. Ex: "precip_adj_\%s\%s\%s.nc"}

\item{obs.hour}{observation hour in local time. Default 8h for Rwanda}

\item{min.frac}{minimum fraction of non missing values}

\item{longname}{long name for the NetCDF variable precip}
}
\description{
Aggregate hourly precipitation to daily.
}
