% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipAccumulation.R
\name{qpe_accumulation}
\alias{qpe_accumulation}
\title{Precipitation accumulation}
\usage{
qpe_accumulation(tstep, time, accumul, min_frac, dirQPE)
}
\arguments{
\item{tstep}{time step of the data: "hourly" or "daily"}

\item{time}{the time/date at which the total will be computed}

\item{accumul}{accumulation duration}

\item{min_frac}{minimum fraction of non missing values}

\item{dirQPE}{full path to the folder containing the qpe data}
}
\description{
Compute precipitation accumulation for a given duration.
}
