% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadUserData.R
\name{readUploadCsvCoords}
\alias{readUploadCsvCoords}
\title{Read csv file}
\usage{
readUploadCsvCoords(file_csv)
}
\arguments{
\item{file_csv}{full path to the csv file}
}
\value{
A json object
}
\description{
Read csv file uploaded by the user.
}
\details{
The data must contain at least 3 column in the order "Name", "Longitude" and "Latitude",
and have a header.
}
