% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadUserData.R
\name{readUploadShptoGeoJSON}
\alias{readUploadShptoGeoJSON}
\title{Read shapefiles}
\usage{
readUploadShptoGeoJSON(dsn, layer)
}
\arguments{
\item{dsn}{full path to the the folder containing files}

\item{layer}{the files name without extension}
}
\value{
A geojson object
}
\description{
Read shapefiles uploaded by the user.
}
\details{
Require the 4 files (.shp, .shx, .dbf and .prj)
}
