% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationStats.R
\name{stats_continuous}
\alias{stats_continuous}
\title{Compute continuous statistics.}
\usage{
stats_continuous(x, y)
}
\arguments{
\item{x}{matrix of the reference observations data used to validate the simulated or estimated data.\cr
The row represents the observations \cr
The column represents the points or stations}

\item{y}{matrix, same dimension as \code{x}, containing the simulated or estimated data to validate.}
}
\value{
A list object
\itemize{
\item{\strong{statistics}: }{a matrix containing the values of the computed statistics.\cr
                    The column represents the points/stations and the row for the statistics}
\item{\strong{description}: }{vector describing the statistics}
\item{\strong{perfect.score}: }{vector of the perfect score values for each statistics}
}
}
\description{
Compute continuous statistics for validation.
}
\references{
\url{https://www.cawcr.gov.au/projects/verification/}
}
