% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsHourlyPrecipitation.R
\name{awsGetHourlyPrecip}
\alias{awsGetHourlyPrecip}
\title{Get hourly precipitation data.}
\usage{
awsGetHourlyPrecip(start_time, end_time, url)
}
\arguments{
\item{start_time}{start time, format "YYYY-mm-dd HH:00"}

\item{end_time}{end time, format "YYYY-mm-dd HH:00"}

\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}
}
\value{
a list object
}
\description{
Get hourly precipitation data for all AWS.
}
