% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeQPECAPPI.R
\name{computeQPECAPPI}
\alias{computeQPECAPPI}
\title{Compute QPE from CAPPI.}
\usage{
computeQPECAPPI(
  url,
  dirOUT,
  start_time,
  end_time,
  cappi = list(method = "composite_altitude", pars = list(fun = "maximum", min_alt =
    1.7, max_alt = 15)),
  qpe = list(method = "RATE_Z", pars = list(alpha = 300, beta = 1.4, invCoef = FALSE)),
  dbz_thres = list(min = 20, max = 65),
  apply_cmd = TRUE,
  pia = NULL,
  filter = NULL
)
}
\arguments{
\item{url}{the URL of the server. Ex: "http://192.168.1.10:8080"}

\item{dirOUT}{full path to the directory to save the output}

\item{start_time}{start time (in local time), format "yyyy-mm-dd HH:MM"}

\item{end_time}{end time (in local time), format "yyyy-mm-dd HH:MM"}

\item{cappi}{a list of parameters to be used to create the CAPPI}

\item{qpe}{a list of parameters to be used to compute the precipitation rate}

\item{dbz_thres}{a list of the minimum and maximum reflectivity threshold}

\item{apply_cmd}{logical, apply clutter mitigation decision}

\item{pia}{a list of parameters to be used to correct the attenuation. NULL no correction performed.}

\item{filter}{a list of parameters to be used to filter the data. NULL no filtering applied.}
}
\description{
Compute QPE from CAPPI for single scan radar polar data.
}
