% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRadarGrid.R
\name{extractRadarGrid}
\alias{extractRadarGrid}
\title{Extract radar Cartesian data.}
\usage{
extractRadarGrid(
  url,
  source,
  start_time,
  end_time,
  fields,
  points,
  levels = -1,
  padxyz = c(0, 0, 0),
  fun_sp = "mean",
  time_zone = "Africa/Kigali"
)
}
\arguments{
\item{url}{The URL of the server. Ex: "http://192.168.1.10:8080"}

\item{source}{A partial path of the directory containing the folders dates of the mdv files.
Ex: "radarCart/ops", "ctrec" or "radarCart/echo_tops"}

\item{start_time, end_time}{The start and end time same time zone as \code{time_zone}, format "YYYY-mm-dd HH:MM"}

\item{fields}{Vector of the fields to extract.}

\item{points}{A data frame of the points to extract. Data frame with column names "id", "longitude" and "latitude"}

\item{levels}{A vector of the index of altitudes to be extracted in integer, or -1 to extract all available altitude}

\item{padxyz}{A vector of the padding to use in number of pixels, in order "lon", "lat", "alt"}

\item{fun_sp}{Character, function to be used for the padding. Options: "mean", "median", "max", "min"}

\item{time_zone}{the time zone of \code{start_time}, \code{end_time} and the output QPE.
Options: "Africa/Kigali" or "UTC". Default "Africa/Kigali"}
}
\value{
A named list
  \itemize{
    \item{\strong{points}: }{the original points used to extract the data}
    \item{\strong{date}: }{list of dates of the extracted data}
    \item{\strong{altitude}: }{list of the altitudes of the extracted data}
    \item{\strong{data}: }{list of the fields in the form of 3d array. \cr
            Dimension: (length(date) x length(altitude) x length(points))}
  }
}
\description{
Extract radar Cartesian data over a set of points.
}
